% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_network_attr.R
\name{cal_network_attr}
\alias{cal_network_attr}
\title{Calculate network topological property for each network}
\usage{
cal_network_attr(network_list)
}
\arguments{
\item{network_list}{a list with multiple networks; all the networks should be trans_network object created from \code{\link{trans_network}} class of \code{microeco} package.}
}
\value{
\code{data.frame}
}
\description{
Calculate the topological properties of all the networks and merge the results into one table.
}
\examples{
data(soil_amp_network)
test <- cal_network_attr(soil_amp_network)

}
