% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{check_input}
\alias{check_input}
\title{Check the input data format}
\usage{
check_input(input = NULL)
}
\arguments{
\item{input}{default NULL; character or data.frame; matching table used.}
}
\description{
Check the input data format
}
