% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_comp.R
\name{edge_comp}
\alias{edge_comp}
\title{Generate a \code{microtable} object with paired nodes distributions of edges across networks}
\usage{
edge_comp(network_list)
}
\arguments{
\item{network_list}{a list with multiple networks; all the networks should be trans_network object created from \code{trans_network} class of microeco package.}
}
\value{
\code{microtable} object
}
\description{
Generate a \code{microtable} object with paired nodes distributions of edges across networks. Useful for the edge comparisons across different networks.
The return \code{otu_table} in \code{microtable} object has the binary numbers in which 1 represents the presence of the edge in the corresponding network.
}
\examples{
data(soil_amp_network)
test <- edge_comp(soil_amp_network)
# test is a microtable object

}
