% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_edge_table.R
\name{get_edge_table}
\alias{get_edge_table}
\title{Get edge property tables from networks}
\usage{
get_edge_table(network_list)
}
\arguments{
\item{network_list}{a list with multiple networks; all the networks should be trans_network object created from \code{\link{trans_network}} class of microeco package.}
}
\value{
list
}
\description{
Get edge property tables from a list with multiple networks.
}
\examples{
\dontrun{
data(soil_amp_network)
soil_amp_network \%<>\% get_edge_table
}
}
