% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_node_table.R
\name{get_node_table}
\alias{get_node_table}
\title{Get node property tables from networks}
\usage{
get_node_table(network_list, ...)
}
\arguments{
\item{network_list}{a list with multiple networks; all the networks should be trans_network object created from \code{\link{trans_network}} class of microeco package.}

\item{...}{parameter passed to get_node_table function of \code{\link{trans_network}} class.}
}
\value{
list
}
\description{
Get node property tables from a list with multiple networks.
}
\examples{
\dontrun{
data(soil_amp_network)
soil_amp_network \%<>\% get_node_table(node_roles = FALSE)
}
}
