% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{meconetcomp}
\alias{meconetcomp}
\title{Introduction to meconetcomp package (\href{https://github.com/ChiLiubio/meconetcomp}{https://github.com/ChiLiubio/meconetcomp})}
\description{
For the detailed tutorial on meconetcomp package, please follow the links:\cr
Online tutorial website: \href{https://chiliubio.github.io/microeco_tutorial/meconetcomp-package.html}{https://chiliubio.github.io/microeco_tutorial/meconetcomp-package.html} \cr
Download tutorial: \href{https://github.com/ChiLiubio/microeco_tutorial/releases}{https://github.com/ChiLiubio/microeco_tutorial/releases}

Please open the help document by using \code{help} function or by clicking the following links collected:\cr
\code{\link{cal_module}}\cr
\code{\link{cal_network_attr}}\cr
\code{\link{get_node_table}}\cr
\code{\link{get_edge_table}}\cr
\code{\link{node_comp}}\cr
\code{\link{edge_comp}}\cr
\code{\link{edge_node_distance}}\cr
\code{\link{edge_tax_comp}}\cr
\code{\link{subset_network}}\cr
\code{\link{subnet_property}}\cr
\code{\link{robustness}}\cr

To report bugs or discuss questions, please use Github Issues (\href{https://github.com/ChiLiubio/meconetcomp/issues}{https://github.com/ChiLiubio/meconetcomp/issues}).
Before creating a new issue, 
please read the guideline (\href{https://chiliubio.github.io/microeco_tutorial/notes.html#github-issues}{https://chiliubio.github.io/microeco_tutorial/notes.html#github-issues}).

To cite meconetcomp package in publications, please run the following command to get the reference: \code{citation("meconetcomp")}\cr

Reference:\cr
Chi Liu, Chaonan Li, Yanqiong Jiang, Raymond Jianxiong Zeng, Minjie Yao, and Xiangzhen Li. 2023.
A guide for comparing microbial co-occurrence networks. iMeta. 2(1): e71.
}
\keyword{Description}
