% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_comp.R
\name{node_comp}
\alias{node_comp}
\title{Generate a microtable object with node distributions across networks}
\usage{
node_comp(network_list, degree = 1)
}
\arguments{
\item{network_list}{a list with multiple networks; all the networks should be trans_network object created from \code{\link{trans_network}} class of microeco package.}

\item{degree}{default 1; the threshold of degree of nodes remained in the network.}
}
\value{
microtable object
}
\description{
Generate a microtable object with node distributions across networks. Useful for the node information comparisons across different networks.
The return otu_table in microtable object has the binary number that 1 represent presence of the node in the corresponding network.
}
\examples{
\dontrun{
data(soil_amp_network)
test <- node_comp(soil_amp_network)
# test is a microtable object
}
}
