% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustness.R
\name{robustness}
\alias{robustness}
\title{Calculate robustness across networks.}
\description{
This class is a wrapper for robustness calculation and the visualization.
}
\examples{

## ------------------------------------------------
## Method `robustness$plot`
## ------------------------------------------------

\donttest{
t1$plot(boxplot_add = "none", add_sig = TRUE)
}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-robustness-new}{\code{robustness$new()}}
\item \href{#method-robustness-plot}{\code{robustness$plot()}}
\item \href{#method-robustness-clone}{\code{robustness$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-robustness-new"></a>}}
\if{latex}{\out{\hypertarget{method-robustness-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{robustness$new(
  network_list,
  link_remove_strategy = c("Rand")[1],
  link_remove_perc = seq(0, 1, 0.1),
  measure = c("Eff")[1],
  run = 10
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{network_list}}{a list with multiple networks; all the networks should be \code{trans_network} object 
created from \code{\link{trans_network}} class of microeco package.}

\item{\code{link_remove_strategy}}{default "Rand"; 
\describe{
   item{\strong{"Rand"}} {links are randomly removed.}
}}

\item{\code{link_remove_perc}}{default seq(0, 1, 0.1).}

\item{\code{measure}}{default "Eff"; the network functioning measures as the representatives of robustness. 
\describe{
   item{\strong{"Eff"}} {network efficiency.}
}}

\item{\code{run}}{default 10.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{res_table}, stored in the object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-robustness-plot"></a>}}
\if{latex}{\out{\hypertarget{method-robustness-plot}{}}}
\subsection{Method \code{plot()}}{
Plot the distance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{robustness$plot(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{parameters pass to \code{plot_alpha} function of \code{trans_alpha} class of \code{microeco} package.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{ggplot}.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\donttest{
t1$plot(boxplot_add = "none", add_sig = TRUE)
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-robustness-clone"></a>}}
\if{latex}{\out{\hypertarget{method-robustness-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{robustness$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
