% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_network.R
\name{save_network}
\alias{save_network}
\title{Save networks as gexf-style files}
\usage{
save_network(network_list, dirpath = ".")
}
\arguments{
\item{network_list}{a list with multiple networks; all the networks should be trans_network object created from \code{\link{trans_network}} class of microeco package.}

\item{dirpath}{default "."; the directory used to save all the gexf files; the default "." represents the current working directory.}
}
\value{
NULL.
}
\description{
Save all networks as gexf-style files, which can be opened by Gephi (\href{https://gephi.org/}{https://gephi.org/}).
}
