% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{soil_amp}
\alias{soil_amp}
\title{The soil_amp data}
\usage{
data(soil_amp)
}
\description{
The soil_amp data is the 16S rRNA gene amplicon sequencing dataset of Chinese wetland soil.
}
\keyword{microtable}
