% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{soil_amp_network}
\alias{soil_amp_network}
\title{The soil_amp_network data}
\usage{
data(soil_amp_network)
}
\description{
The soil_amp_network data is a list storing three trans_network objects created based on soil_amp data.
Three networks are created for IW, CW and TW groups, respectively.
}
\keyword{list}
