% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{soil_measure_diversity}
\alias{soil_measure_diversity}
\title{The soil_measure_diversity data}
\usage{
data(soil_measure_diversity)
}
\description{
The soil_measure_diversity data is a table storing all the abiotic factors and functional diversity based on the metagenomic sequencing and MetaCyc pathway analysis.
}
\keyword{data.frame}
