% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stool_met}
\alias{stool_met}
\title{The stool_met data}
\usage{
data(stool_met)
}
\description{
The stool_met data is the metagenomic species abundance dataset of stool samples selected from R ExperimentHub package.
It has 198 samples, collected from the people with alcohol drinking habit, and 92 species.
}
\keyword{microtable}
