% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subnet_property.R
\name{subnet_property}
\alias{subnet_property}
\title{Calculate properties of sub-networks selected according to features in samples}
\usage{
subnet_property(network_list)
}
\arguments{
\item{network_list}{a list with multiple networks; all the networks should be trans_network object created from \code{trans_network} class of \code{microeco} package.}
}
\value{
\code{data.frame}
}
\description{
Extracting sub-network according to the presence of features in each sample across networks and calculate the sub-network properties.
}
\examples{
\donttest{
data(soil_amp_network)
test <- subnet_property(soil_amp_network)
}
}
