% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_network.R
\name{subset_network}
\alias{subset_network}
\title{Extract subset of network according to the edge intersection of networks}
\usage{
subset_network(network_list, venn = NULL, name = NULL)
}
\arguments{
\item{network_list}{a list with multiple networks; all the networks should be trans_network object created from \code{\link{trans_network}} class of microeco package.}
}
\value{
a trans_network object, with the extracted edges in the network
}
\description{
Extracting a network according to the edge intersection of networks.
}
\examples{
\dontrun{
data(soil_amp_network)
# obtain edge distribution
tmp <- edge_comp(soil_amp_network)
# obtain edge intersection using trans_venn class
library(microeco)
tmp1 <- trans_venn$new(tmp)
# convert intersection result to microtable object
tmp2 <- tmp1$trans_comm()
# extract all the intersection of "IW", "TW" and "CW"
IWTWCW_intersec <- subset_network(soil_amp_network, venn = tmp2, name = "IW&TW&CW")
IWTWCW_intersec trans_network object
}
}
