% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustness.R
\name{robustness}
\alias{robustness}
\title{Calculate robustness across networks.}
\description{
This class is a wrapper for robustness calculation and visualization.
}
\examples{

## ------------------------------------------------
## Method `robustness$plot`
## ------------------------------------------------

\donttest{
t1$plot()
}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-robustness-new}{\code{robustness$new()}}
\item \href{#method-robustness-plot}{\code{robustness$plot()}}
\item \href{#method-robustness-clone}{\code{robustness$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-robustness-new"></a>}}
\if{latex}{\out{\hypertarget{method-robustness-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{robustness$new(
  network_list,
  remove_strategy = c("edge_rand", "edge_strong", "edge_weak", "node_rand", "node_hub",
    "node_degree_high", "node_degree_low")[1],
  remove_ratio = seq(0, 1, 0.1),
  measure = c("Eff", "Eigen")[1],
  run = 10
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{network_list}}{a list with multiple networks; all the networks should be \code{trans_network} object 
created from \code{\link{trans_network}} class of microeco package.}

\item{\code{remove_strategy}}{default "edge_rand"; 
\describe{
     \item{\strong{"edge_rand"}}{edges are randomly removed.}
  \item{\strong{"edge_strong"}}{edges are removed in decreasing order of weight.}
     \item{\strong{"edge_weak"}}{edges are removed in increasing order of weight.}
     \item{\strong{"node_rand"}}{nodes are removed randomly.}
     \item{\strong{"node_hub"}}{node hubs are removed. The hubs include network hubs and module hubs.}
     \item{\strong{"node_degree_high"}}{nodes are removed in decreasing order of degree.}
     \item{\strong{"node_degree_low"}}{nodes are removed in increasing order of degree.}
}}

\item{\code{remove_ratio}}{default seq(0, 1, 0.1).}

\item{\code{measure}}{default "Eff"; the network functioning measures as the representatives of robustness. 
\describe{
     \item{\strong{"Eff"}}{network efficiency. The average efficiency of the network is defined:
           \deqn{Eff = \frac{1}{N(N - 1)} \sum_{i \neq j \in G}\frac{1}{d(i, j)}}
          where N is the total number of nodes and \emph{d(i,j)} is the shortest path between node i and node j. 
          When the weight is found in the edge attributes, \eqn{d(i,j)} denotes the weighted shortest path between node i and node j.
          }
     \item{\strong{"Eigen"}}{natural connectivity, represented as an average eigenvalue that changes strictly monotonically with the addition or deletion of edges.}        
}}

\item{\code{run}}{default 10. Replication number applied for the sampling method.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{res_table}, stored in the object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-robustness-plot"></a>}}
\if{latex}{\out{\hypertarget{method-robustness-plot}{}}}
\subsection{Method \code{plot()}}{
Plot the simulation results.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{robustness$plot(
  color_values = RColorBrewer::brewer.pal(8, "Dark2"),
  show_point = TRUE,
  point_size = 1,
  point_alpha = 0.6,
  show_errorbar = TRUE,
  errorbar_position = position_dodge(0),
  errorbar_size = 1,
  errorbar_width = 0.1,
  add_fitting = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{color_values}}{colors used for presentation.}

\item{\code{show_point}}{default TRUE; whether show the point.}

\item{\code{point_size}}{default .3; point size value.}

\item{\code{point_alpha}}{default .6; point alpha value.}

\item{\code{show_errorbar}}{default TRUE; whether show the errorbar by using the SD result.}

\item{\code{errorbar_position}}{default position_dodge(0); Position adjustment, either as a string (such as "identity"), 
or the result of a call to a position adjustment function.}

\item{\code{errorbar_size}}{default 1; errorbar size.}

\item{\code{errorbar_width}}{default 0.1; errorbar width.}

\item{\code{add_fitting}}{default FALSE; whether add fitted line.}

\item{\code{...}}{parameters pass to ggplot2::geom_line (when add_fitting = FALSE) or ggplot2::geom_smooth (when add_fitting = TRUE).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{ggplot}.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\donttest{
t1$plot()
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-robustness-clone"></a>}}
\if{latex}{\out{\hypertarget{method-robustness-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{robustness$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
