% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{soil_amp}
\alias{soil_amp}
\title{The soil_amp data}
\usage{
data(soil_amp)
}
\description{
The soil_amp data is the 16S rRNA gene amplicon sequencing dataset of Chinese wetland soils.
Reference: An et al. 2019 <doi:10.1016/j.geoderma.2018.09.035>; Liu et al. 2022 <10.1016/j.geoderma.2022.115866>
}
\keyword{microtable}
